api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Yoshioka <sss.yoshioka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
django-bootstrap5 = "^23.3"
psycopg2 = "^2.9.7"
graphene-django = "^3.1.5"
django-cors-headers = "^4.2.0"
pyjwt = "^2.8.0"
requests = "^2.31.0"
cryptography = "^41.0.3"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.286"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 160
